Move notebook conversion files to their own folder.
